% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add.R
\name{Add}
\alias{Add}
\title{In the simulations, add the mean structure and the noise structure to the original co-variance stucture}
\usage{
Add(inidata, cluster_assign, cluster_num, prob, p_noise)
}
\arguments{
\item{inidata}{The generated data that includes cluster differences in co-variance structure (i.e. loading matices).
The input should be a matrix with the number of rows should be equal to the number of entries \code{N}
while the number of columns should be equal to the total amount of variables \code{p} (over all blocks).}

\item{cluster_assign}{A vector indicates the assignment of each entry to one of the cluster. The vector
should be of length \code{N}, and the nth element refers to the number of cluster that the nth entry belong to.}

\item{cluster_num}{A non-negive interger indicates the total number of clusters that would be generated.}

\item{prob}{A number within the range of [0,1] that indicates the proportion of mean-level differences
in the total cluster difference; the proportion of total difference that is accounted by co-variance difference
would therefore be \code(1 - prob).}

\item{p_noise}{A number within the range of [0,1] that indicates the percentage of noise structrue that should be added to the final data.}
}
\value{
a list of two elements. The first element is the generated final data and the second element is the noise structure
}
\description{
In the simulations, add the mean structure and the noise structure to the original co-variance stucture
}
\examples{
m <- matrix(1:6,2,3) #N = 2
n_cluster <- 2
cluster_assignment <- c(2,1)
p_mean <- 0.9 #90\% mean structure
p_noise <- 0.1 # 10\% noise
Add(m, n_cluster, cluster_assignment, p_mean, p_noise)
}
