% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListCongruence.R
\name{ListCongruence}
\alias{ListCongruence}
\title{Calculate the Tucker Congruence of two lists of matrices}
\usage{
ListCongruence(lista, listb, cluster_num)
}
\arguments{
\item{lista}{the first list with matrix elements}

\item{listb}{the second list with matrix elements}

\item{cluster_num}{the number of cluster (which should also be the maximum length of the two lists)}
}
\value{
a vector of length cluster_num. \code{nth} elemen of the vector indicates the optimal congruence between
the \code{nth} matrix in lista and the listb (maximum possible value)
}
\description{
Calculate the Tucker Congruence of two lists of matrices
}
\examples{
# calculate the simulated (original) data and the (re-covered) data
(not run)  sim <- CSSCASimulation(n_cluster, mem_cluster, n_block, n_com, n_distinct, n_var, p_sparse,
 p_noise, p_combase, p_fixzero, "both", mean_v)
(not run)  results <- MainCSSCA(sim$concatnated_data, n_var, n_block, n_com, n_distinct, n_cluster, n_obs, p_sparse, sim$cluster_assign, n_replace = 5)
# calculate the average congruence between the simulated loading matrices and estimated loading matrices
ListCongruence(sim$loading, results$loading)

}
