% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuckerCongruence.R
\name{TuckerCongruence}
\alias{TuckerCongruence}
\title{Calculate the Tucker Congruence of two matrices which allow for rotation freedom.
Note that the factor.congruence in the "psych" package would not allow for rotation freedom}
\usage{
TuckerCongruence(matrix1, matrix2)
}
\arguments{
\item{matrix1}{the first matrix}

\item{matrix2}{the second matrix}
}
\value{
the congruence between the two matrices which is in theory
the maximum possible congruence taking into account all possible rotations
}
\description{
Calculate the Tucker Congruence of two matrices which allow for rotation freedom.
Note that the factor.congruence in the "psych" package would not allow for rotation freedom
}
\examples{

}
